Extract MemberInfo component
