+++ packages/ui/src/pages/Profile/MyMemberships/MemberItem.tsx
@@ -1,8 +1,7 @@
-import React, { useState } from 'react'
-import styled from 'styled-components'
+import React from 'react'
 import { ButtonGhostMediumSquare } from '../../../components/buttons'
+import { MemberInfo } from '../../../components/MemberInfo'
 import { Member } from '../../../modals/AddMembershipModal/MembershipFormModal'
-import { MembershipAbout } from './MembershipAbout'
 
 interface MemberProps {
   member: Member
@@ -16,28 +15,3 @@ export const MemberItem = ({ member }: MemberProps) => {
     </div>
   )
 }
-
-export const MemberInfo = ({ member }: MemberProps) => {
-  const [isAboutOpen, setAboutOpen] = useState(false)
-
-  return (
-    <>
-      <div>
-        <Avatar src={member.avatarURI} />
-        <MemberHandle onClick={() => setAboutOpen(true)}>{member.handle}</MemberHandle>
-        <p>0 times</p>
-      </div>
-      {isAboutOpen && <MembershipAbout member={member} onClose={() => setAboutOpen(false)} />}
-    </>
-  )
-}
-
-export const Avatar = styled.img`
-  border-radius: 50%;
-`
-
-const MemberHandle = styled.a`
-  font-weight: bold;
-  text-decoration: underline;
-  cursor: pointer;
-`
